requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
langchain-core
# Core ML/AI
openai>=1.0.0
anthropic>=0.8.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-anthropic>=0.1.0

# LLM & Fine-tuning (GPU support)
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=3.20.0
scipy>=1.11.0

# HuggingFace
huggingface-hub>=0.19.0
langgraph
tiktoken
numpy
pandas
faiss-cpu
ruff
pytest
pytest-asyncio
typer
rich
transformers>=4.30.0
huggingface-hub
accelerate
peft
bitsandbytes
datasets
torch